Remove sign warning

+++ PixelsGenerator.cpp
@@ -11,10 +11,10 @@ template <typename T>
 T* generateRandomPixels(GLushort width, GLushort height, GLubyte depth)
 {
 	srand((int) time(NULL));
-	
+
 	T* pixels = (T*) calloc(width * height * depth, sizeof(T));
-	for (GLuint i = 0; i < width * height * depth; i++)
+	for (GLuint i = 0; i < (GLuint) width * height * depth; i++)
 		pixels[i] = rand() % (T) (ceil(pow(2, 8 * sizeof(T))) - 1);
-		
+
 	return pixels;
 }
